Be more explicit in pulling clusters from seurat metadata file
First, check if this is the metadata from the seurat object by checking
the first two header columns. The first must be blank, and the second
must be `orig.ident`. Once that is confirmed, look for a column named
`X__dredge_cluster`. This will be the column used to assign a cell to a
cluster.