Draw selected transcripts in SC HeatMap
As part of having a variable number of transcripts in the table, I
started drawing the hovered transcript in an independent canvas. When
things were too bunched together, the transcript label for the hovered
would bleed out of the rectangle it was in. Putting it in a separate
canvas means we can just clear the whole thing when a new one is
hovered.

I also made two additional tweaks. 1) There is a maximum height for the
boxes in the heatmap grid, and 2) the transcript label font scales down
as the rows in the heatmap get shorter.