Make some changes to GenericTable to help out type checker
The new list components in react-window use a generic type to add type
information to the `itemData` prop. As the code was, TypeScript wasn't
inferring the type of that parameter during my calls to
React.createElement. This commit adds a number of explicit type
annotations to help it out.

I think this might not have been an issue if I were using JSX, but I am
not definitely sure.