Don't require holding ctrl to select multiple transcripts
Also, don't draw border around selected transcripts. It looked weird
when there were multiple in a row.

This also fixes a bug when selecting the next focused transcript after
removing a selected transcript (i.e. the transcript that's displayed in
the little in-situ bar). What we want to do is to create an array from
the newly constructed set of selected transcripts, take one transcript
off the end (if it exists), and set that as the next focused transcript.
However, this was constructing an array from the focused transcript
*name* (rather than the set of selected transcripts). So it was setting
the next focused transcript as the last letter of the transcript that
was clicked. This probably accounts for sporadic "this transcript was
not found in the dataset" errors!