Fix a bug in loading transcripts
In JS, objects preserve insertion order, UNLESS the keys are integers.
(All keys in objects are strings, so actually, if the key parses to an
integer). There was a weird bug where I was relying on insertion order
for keeping track of transcripts, but one of the transcripts was named
"41". This commit switches to using a Map rather than an object for
tracking the transcripts.

This commit also index transcripts that are not differentially expressed
in any cluster.